backend/pyproject.toml
[project]
name = "brocodde-backend"
version = "0.4.0"
description = "BroCoDDE — Content Development Life Cycle Engine (Backend)"
authors = [{ name = "Prabakaran Chandran", email = "prabakaran@prachalabs.com" }]
requires-python = ">=3.12"

dependencies = [
    # Agent runtime / frameworks
    "agno>=1.0.0",
    "pydantic-ai>=0.0.36",
    "pydantic-ai-slim[anthropic,openai,groq]>=0.0.36",

    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1.0",

    # HTTP client
    "httpx>=0.27.0",
    "exa-py>=1.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",

    # AI providers
    "openai>=1.50.0",  # Whisper transcription
    "anthropic>=0.34.0",

    # Config / utilities
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "python-slugify>=8.0.0",
    "tenacity>=8.5.0",
    "rich>=13.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
